atarus missile
Any Legio Atarus Titan armed with Apocalypse missile
launchers, Apocalypse missile arrays or an Apocalypse missile
launcher can be upgraded to have Infernus missiles for +25
points, so long as that Titan does not already have an upgrade
affecting that weapon. Each of the Titan’s weapons must be
upgraded separately. A weapon with this upgrade gains the
Voidbreaker (2) trait. In addition, if a unit suffers at least one
hit from a weapon with this upgrade, place a 5" Blast marker
under the unit. In the End phase, any model with at least part
of its base on the marker takes a Strength 4 hit, ignoring void
shields and ion shields. After resolving any hits, the fire goes
out and the marker is removed. Titans suffer this hit to their
Legs. In addition, Blocking terrain touched by the marker takes
a Strength 4 hit (see the Adeptus Titanicus rulebook for details
on damaging and destroying terrain) and, if not destroyed, is set
alight. In the End phase, burning terrain suffers a Strength 4 hit,
and deals a Strength 4 hit to any models in it, as detailed above. The fire then goes out.